added new redux config refactor
