+++ package.json
@@ -42,6 +42,7 @@
     "@angular/router": "^8.2.1",
     "@ecodev/fab-speed-dial": "4.0.0",
     "@ngrx/store": "^8.5.1",
+    "@ngrx/store-devtools": "^8.5.1",
     "@ngtools/webpack": "^8.2.1",
     "@ngx-pwa/local-storage": "^5.3.0",
     "@ngx-translate/core": "^11.0.1",
