+++ src/app/redux/reducers/layer-filter-reducer.spec.ts
@@ -1,0 +1,12 @@
+import * as layerFilterReducer from './layer-filter-reducer'
+
+describe.only('My Selectors', () => {
+  it('should calc selectTotal', () => {
+    console.log('====================================')
+    console.log(layerFilterReducer)
+    console.log('====================================')
+    // expect(layerFilterReducer.selectTotal.projector(2, 3)).toBe(5);
+  })
+})
+
+
